requirements.txt
spotipy==2.23.0
gitpython==3.1.40
vaderSentiment==3.3.2
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
